requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# NLP and text processing
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
networkx>=3.1

# Machine learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Data storage and processing
json5>=0.9.14

# Web framework for dashboard
streamlit>=1.25.0
dash>=2.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
flake8>=6.0.0

# Additional utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
